refactor(android): extract Health pure helpers
