+++ android/app/src/main/java/com/noop/ui/HealthScreen.kt
@@ -1125,87 +1125,6 @@ fun VitalSignsScreen(vm: AppViewModel, onVitalClick: (String) -> Unit = {}) {
     }
 }
 
-// MARK: - Derived live HR
-//
-// HR to display: the reported value when > 0, else derived from the latest R-R
-// interval in milliseconds (the strap streams R-R even when its HR field reads 0).
-
-private fun displayHr(bpm: Int?, live: LiveState): Int? {
-    // #39: prefer the spike-filtered median (AppViewModel.bpm) over raw live.heartRate, which carries
-    // PPG harmonic spikes (real ~92 read as 170+). Raw / R-R are last-resort fallbacks.
-    if (bpm != null && bpm > 0) return bpm
-    live.heartRate?.let { if (it > 0) return it }
-    val lastRr = live.rr.lastOrNull()
-    if (lastRr != null && lastRr > 0) return (60_000.0 / lastRr).roundToInt()
-    return null
-}
-
-private fun hrIsDerived(live: LiveState): Boolean =
-    (live.heartRate ?: 0) <= 0 && live.rr.isNotEmpty()
-
-/** HR as a fraction of HR-max (0..1). */
-private fun hrFraction(hr: Int?, hrMax: Int): Double {
-    if (hr == null || hrMax <= 0) return 0.0
-    return (hr.toDouble() / hrMax).coerceIn(0.0, 1.0)
-}
-
-/** Current zone 1..5 from %HR-max (WHOOP/Karvonen-style bands: 50/60/70/80/90). */
-private fun hrZone(fraction: Double): Int = when {
-    fraction < 0.60 -> 1
-    fraction < 0.70 -> 2
-    fraction < 0.80 -> 3
-    fraction < 0.90 -> 4
-    else -> 5
-}
-
-/** One streamed live-HR reading with the wall-clock time it arrived (epoch millis). Carrying the
- *  time — not a bare bpm — is what lets the hero render a real time x-axis (#198). */
-data class LiveHrSample(val timeMs: Long, val bpm: Double)
-
-/** A short, time-stamped HR series for the hero chart. Prefers the accumulated live-HR history
- *  (which moves over time); falls back to per-beat HR from R-R, then to a flat pair while the
- *  buffer fills. The old version derived ONLY from R-R, which is sparse on WHOOP 4, so it sat on a
- *  flat 2-point line even while HR was clearly changing (issue #18). The R-R / flat fallbacks have
- *  no real per-sample timestamps, so we synthesise a 1 Hz trailing window ending "now" — the x-axis
- *  still reads as clock time and scrolls, matching the live buffer (#198). */
-private fun hrSeries(history: List<LiveHrSample>, live: LiveState, hr: Int?): List<LiveHrSample> {
-    if (history.size > 1) return history
-    val beats = live.rr.takeLast(60).mapNotNull { rr ->
-        if (rr > 0) 60_000.0 / rr else null
-    }
-    if (beats.size > 1) return synthesiseSeries(beats)
-    if (hr != null) return synthesiseSeries(listOf(hr.toDouble(), hr.toDouble()))
-    return emptyList()
-}
-
-/** Wrap a bare value series in trailing 1 Hz timestamps ending "now", so the fallbacks chart on the
- *  same time x-axis as the live buffer. */
-private fun synthesiseSeries(values: List<Double>): List<LiveHrSample> {
-    val now = System.currentTimeMillis()
-    val n = values.size
-    return values.mapIndexed { i, v ->
-        LiveHrSample(timeMs = now + (i - (n - 1)) * 1000L, bpm = v)
-    }
-}
-
-/** The live-HR hero's rolling buffer cap: 180 samples at the 1 Hz tick (#941) is a strict ~3 minutes. */
-internal const val LIVE_HR_BUFFER_CAP = 180
-
-/** One 1 Hz tick of the hero buffer (#941): bank the latest smoothed HR when it is present and
- *  physiologically plausible (30..220, the same range guard the old on-change append used), then trim
- *  the buffer to the rolling cap. Pure so the guard + cap behaviour is JVM-testable. */
-internal fun appendLiveHrSample(
-    history: MutableList<LiveHrSample>,
-    bpm: Int?,
-    timeMs: Long,
-    cap: Int = LIVE_HR_BUFFER_CAP,
-) {
-    val v = bpm ?: return
-    if (v !in 30..220) return
-    history.add(LiveHrSample(timeMs = timeMs, bpm = v.toDouble()))
-    while (history.size > cap) history.removeAt(0)
-}
-
 // MARK: - Heart rate hero (live)
 
 @Composable
@@ -1581,241 +1500,6 @@ private fun VitalsSection(
 
 // MARK: - Vital model
 
-private data class Vital(
-    val key: String,
-    val label: String,
-    val unit: String,
-    val value: Double?,
-    val format: (Double) -> String,
-    val deltaText: String? = null,
-    val readingDay: String? = null,
-    val asOfLabel: String? = null,
-    val rangeCaption: String? = null,
-    /** Metric-specific "no value" line, shown in place of a bare "No data" when nothing resolved, so an
-     *  empty tile still says WHY. Required (no default) so a new vital must state its own reason.
-     *  Kotlin twin of `BodyVitalReading.missingCaption` (VitalSignsSummary.swift) — same wording. */
-    val missingCaption: String,
-    /** Personal-baseline banding (population fallback until 14 trusted nights). */
-    val banding: VitalBands.Result,
-    /** The metric's category colour (used only when in range). */
-    val metricColor: Color,
-    /** Trailing values (oldest → newest) for the tile's metric-tinted sparkline trail, matching
-     *  Today's Key-Metrics tiles. Presentation-only; defaulted so existing call sites compile. */
-    val sparkline: List<Double> = emptyList(),
-) {
-    /** Value with its unit appended, or null when no data. */
-    val formattedValue: String? = value?.let { "${format(it)} $unit" }
-
-    /** Colour communicates state: in-range = the metric's category colour,
-     *  out-of-range = warning amber, no data = tertiary. */
-    val accent: Color = when (banding.band) {
-        VitalBands.Band.NO_DATA -> Palette.textTertiary
-        VitalBands.Band.IN_RANGE -> metricColor
-        VitalBands.Band.OUT_OF_RANGE -> Palette.statusWarning
-    }
-
-    /** The in-range caption that stands in for a StatePill inside the fixed-height tile.
-     *  The wording says which yardstick judged it: your baseline vs typical ranges. */
-    val stateCaption: String = when {
-        // Raw SpO₂ is a device-dependent ADC, not a clinical value — never claim an in/out-of-range
-        // judgment. Show a plain "uncalibrated" note when a value decoded. (#93)
-        key == "spo2raw" && banding.band != VitalBands.Band.NO_DATA -> "Uncalibrated"
-        // Nothing resolved: say WHY this tile is empty rather than a bare "No data", which reads as a
-        // bug for metrics NOOP cannot derive from a strap at all (the calibrated SpO₂ % is import-only:
-        // AnalyticsEngine writes spo2Pct = null on purpose, see Spo2ReTrace). Ports the Apple behaviour
-        // (`guard let day else { return missingCaption }`) — Android showed "No data" for every case.
-        banding.band == VitalBands.Band.NO_DATA -> missingCaption
-        banding.basis == VitalBands.Basis.PERSONAL ->
-            if (banding.band == VitalBands.Band.IN_RANGE) "In your range" else "Off your baseline"
-        else ->
-            if (banding.band == VitalBands.Band.IN_RANGE) "In typical range" else "Outside typical range"
-    }
-
-    val accessibilityText: String =
-        formattedValue?.let {
-            listOfNotNull("$label: $it", asOfLabel, stateCaption).joinToString(", ")
-        } ?: "$label: no data"
-}
-
-private enum class VitalCaptionMode {
-    AS_OF,
-    RANGE,
-}
-
-/** Build the vitals, banded against the user's OWN trailing baseline once 14 trusted
- *  nights exist (population ranges before that — VitalBands does the deciding). */
-private fun vitalsFor(
-    d: DailyMetric?,
-    days: List<DailyMetric>,
-    tempUnit: TemperatureUnit = TemperatureUnit.CELSIUS,
-): List<Vital> {
-    val todayKey = d?.day
-    // History strictly before the displayed day, oldest→newest (recentDays is already
-    // oldest→newest); calendar-padded so wear gaps count as missing nights (a stale
-    // baseline then falls back to the population range).
-    val history = days.filter { row -> todayKey == null || row.day < todayKey }
-    fun series(selector: (DailyMetric) -> Double?): List<Double?> =
-        VitalBands.calendarSeries(history.map { it.day to selector(it) })
-    fun previous(selector: (DailyMetric) -> Double?): Double? =
-        history.asReversed().asSequence().mapNotNull(selector).firstOrNull()
-    fun deltaText(current: Double?, previous: Double?, decimals: Int = 1): String? {
-        if (current == null || previous == null) return null
-        val diff = current - previous
-        val sign = if (diff >= 0.0) "+" else "-"
-        val mag = kotlin.math.abs(diff)
-        val num = if (decimals == 0) mag.roundToInt().toString()
-        else String.format(Locale.US, "%.${decimals}f", mag)
-        return "($sign$num)"
-    }
-    fun rangeCaption(allValues: List<Double>, unit: String, format: (Double) -> String): String? {
-        val min = allValues.minOrNull() ?: return null
-        val max = allValues.maxOrNull() ?: return null
-        return "within ${format(min)} -- ${format(max)} $unit"
-    }
-    // Trailing values (oldest → newest) feeding each tile's sparkline trail. Built from the same
-    // history already gathered for banding, including the displayed day's value. Presentation-only.
-    fun trail(current: Double?, window: Int = 14, selector: (DailyMetric) -> Double?): List<Double> =
-        (history.mapNotNull(selector) + listOfNotNull(current)).takeLast(window)
-
-    // Skin temp is bimodal: CSV imports store ABSOLUTE °C, the on-device pipeline a ±°C
-    // DEVIATION — partition the history to the displayed value's kind and pick the matching
-    // config + population fallback (±0.6 °C mirrors the illness watch's flag threshold).
-    // This also fixes the live bug where a strap-computed +0.2 °C deviation read
-    // "Out of range" against the 33–36 absolute band.
-    val skin = d?.skinTempDevC
-    // Track which kind the value is so the temperature converter picks the right rule: an ABSOLUTE
-    // reading uses the full C→F formula (×9/5 + 32); a ±DEVIATION must omit the offset.
-    val skinIsAbsolute = skin?.let { VitalBands.isAbsoluteSkinTemp(it) } ?: true
-    val skinResult: VitalBands.Result = if (skin == null) {
-        VitalBands.Result(VitalBands.Band.NO_DATA, VitalBands.Basis.POPULATION, 0)
-    } else {
-        VitalBands.band(
-            value = skin,
-            history = VitalBands.skinTempHistory(skin, series { it.skinTempDevC }),
-            populationRange = if (skinIsAbsolute) 33.0..36.0 else -0.6..0.6,
-            cfg = if (skinIsAbsolute) Baselines.metricCfg.getValue("skin_temp") else VitalBands.skinTempDeviationCfg,
-        )
-    }
-    // Resolve the skin-temp label + converter once, honouring the °C/°F preference. `Vital.formattedValue`
-    // appends `unit`, so strip the trailing " °C/°F" the formatter adds.
-    val skinUnitLabel = UnitFormatter.temperatureUnit(tempUnit)
-    val skinFormat: (Double) -> String = { c ->
-        val full = if (skinIsAbsolute) {
-            UnitFormatter.temperatureFromCelsius(c, tempUnit, decimals = 1)
-        } else {
-            UnitFormatter.temperatureDeltaFromCelsius(c, tempUnit, decimals = 1)
-        }
-        full.removeSuffix(" $skinUnitLabel")
-    }
-    val previousSkin = history.asReversed().asSequence()
-        .mapNotNull { row -> row.skinTempDevC?.takeIf { VitalBands.isAbsoluteSkinTemp(it) == skinIsAbsolute } }
-        .firstOrNull()
-    val respRangeCaption = rangeCaption(days.mapNotNull { it.respRateBpm }, "rpm") { String.format(Locale.US, "%.1f", it) }
-    val spo2RangeCaption = rangeCaption(days.mapNotNull { it.spo2Pct }, "%") { String.format(Locale.US, "%.0f", it) }
-    val rhrRangeCaption = rangeCaption(days.mapNotNull { it.restingHr?.toDouble() }, "bpm") { it.roundToInt().toString() }
-    val hrvRangeCaption = rangeCaption(days.mapNotNull { it.avgHrv }, "ms") { it.roundToInt().toString() }
-    val skinRangeCaption = rangeCaption(
-        days.mapNotNull { row ->
-            row.skinTempDevC?.takeIf { VitalBands.isAbsoluteSkinTemp(it) == skinIsAbsolute }
-        },
-        skinUnitLabel,
-        skinFormat,
-    )
-    // WHOOP 4.0 raw SpO₂: the (red + IR) / 2 ADC mean per night, present only when both channels
-    // decoded for the day. Averaged for a single "signal decoded" tile; both channels stay in the DB. (#93)
-    val spo2RawMean: (DailyMetric) -> Double? = { row ->
-        if (row.spo2Red != null && row.spo2Ir != null) (row.spo2Red + row.spo2Ir) / 2.0 else null
-    }
-    val spo2rawRangeCaption =
-        rangeCaption(days.mapNotNull(spo2RawMean), "ADC") { String.format(Locale.US, "%.0f", it) }
-    return listOf(
-        Vital(
-            key = "resp", label = uiString(R.string.l10n_health_screen_resp_rate_1c48dbd8), unit = "rpm",
-            missingCaption = "No respiratory-rate value",
-            value = d?.respRateBpm, format = { String.format("%.1f", it) },
-            deltaText = deltaText(d?.respRateBpm, previous { it.respRateBpm }),
-            readingDay = todayKey,
-            asOfLabel = asOfLabel(todayKey),
-            rangeCaption = respRangeCaption,
-            banding = VitalBands.band(d?.respRateBpm, series { it.respRateBpm }, 12.0..20.0, Baselines.respCfg),
-            metricColor = Palette.metricCyan,
-            sparkline = trail(d?.respRateBpm) { it.respRateBpm },
-        ),
-        Vital(
-            key = "spo2", label = uiString(R.string.l10n_health_screen_blood_o_9bf5ed9b), unit = "%",
-            missingCaption = "No SpO₂ import or Health value",
-            value = d?.spo2Pct, format = { String.format("%.0f", it) },
-            deltaText = deltaText(d?.spo2Pct, previous { it.spo2Pct }, decimals = 0),
-            readingDay = todayKey,
-            asOfLabel = asOfLabel(todayKey),
-            rangeCaption = spo2RangeCaption,
-            // Population-only on purpose: an absolute <95% floor is meaningful regardless
-            // of personal baseline (no "spo2" MetricCfg exists).
-            banding = VitalBands.band(d?.spo2Pct, emptyList(), 95.0..100.0, null),
-            metricColor = Palette.metricCyan,
-            sparkline = trail(d?.spo2Pct) { it.spo2Pct },
-        ),
-        Vital(
-            // Issue #93: WHOOP 4.0 raw SpO₂ PPG ADC mean (red+IR)/2 per night. NOT a calibrated
-            // blood-oxygen % — that needs WHOOP's proprietary curve. Shown as RAW ADC so users can SEE
-            // the sensor data decoded, without fabricating a clinical-looking number. Banding over the
-            // full u16 span just keeps the tile cyan (never "off range"); `stateCaption` labels it
-            // uncalibrated, so we never assert an in/out-of-range clinical judgment on raw sensor data.
-            key = "spo2raw", label = uiString(R.string.l10n_health_screen_raw_spo_ccfe80c1), unit = "ADC",
-            missingCaption = "No raw SpO₂ decode for the night",
-            value = d?.let(spo2RawMean), format = { String.format("%.0f", it) },
-            deltaText = deltaText(d?.let(spo2RawMean), previous(spo2RawMean), decimals = 0),
-            readingDay = todayKey,
-            asOfLabel = asOfLabel(todayKey),
-            rangeCaption = spo2rawRangeCaption,
-            banding = VitalBands.band(d?.let(spo2RawMean), emptyList(), 0.0..65535.0, null),
-            metricColor = Palette.metricCyan,
-            sparkline = trail(d?.let(spo2RawMean)) { spo2RawMean(it) },
-        ),
-        Vital(
-            key = "rhr", label = uiString(R.string.l10n_health_screen_resting_hr_26677094), unit = "bpm",
-            missingCaption = "No resting HR value",
-            value = d?.restingHr?.toDouble(), format = { it.roundToInt().toString() },
-            deltaText = deltaText(d?.restingHr?.toDouble(), previous { it.restingHr?.toDouble() }, decimals = 0),
-            readingDay = todayKey,
-            asOfLabel = asOfLabel(todayKey),
-            rangeCaption = rhrRangeCaption,
-            banding = VitalBands.band(
-                d?.restingHr?.toDouble(), series { it.restingHr?.toDouble() }, 40.0..60.0,
-                Baselines.restingHRCfg,
-            ),
-            metricColor = Palette.metricRose,
-            sparkline = trail(d?.restingHr?.toDouble()) { it.restingHr?.toDouble() },
-        ),
-        Vital(
-            key = "hrv", label = "HRV", unit = "ms",
-            missingCaption = "No HRV value",
-            value = d?.avgHrv, format = { it.roundToInt().toString() },
-            deltaText = deltaText(d?.avgHrv, previous { it.avgHrv }, decimals = 0),
-            readingDay = todayKey,
-            asOfLabel = asOfLabel(todayKey),
-            rangeCaption = hrvRangeCaption,
-            banding = VitalBands.band(d?.avgHrv, series { it.avgHrv }, 40.0..120.0, Baselines.hrvCfg),
-            metricColor = Palette.metricPurple,
-            sparkline = trail(d?.avgHrv) { it.avgHrv },
-        ),
-        Vital(
-            key = "skin", label = uiString(R.string.l10n_health_screen_skin_temp_a4affc5a), unit = skinUnitLabel,
-            missingCaption = "No nightly skin-temp value",
-            value = skin, format = skinFormat,
-            deltaText = deltaText(skin, previousSkin),
-            readingDay = todayKey,
-            asOfLabel = asOfLabel(todayKey),
-            rangeCaption = skinRangeCaption,
-            banding = skinResult, metricColor = Palette.metricAmber,
-            // Keep the trail on the displayed value's kind — absolute °C and ±deviation must not mix.
-            sparkline = trail(skin) { row ->
-                row.skinTempDevC?.takeIf { VitalBands.isAbsoluteSkinTemp(it) == skinIsAbsolute }
-            },
-        ),
-    )
-}
-
 @Composable
 private fun VitalTile(
     vital: Vital,
@@ -1928,36 +1612,6 @@ private fun TileSparkline(values: List<Double>, color: Color, modifier: Modifier
         drawCircle(color = Palette.tipCore, radius = 1.6f, center = end)
     }
 }
-
-/** One windowed reading behind a vital's detail chart: its day ("YYYY-MM-DD"), the value, and the RAW
- *  source id it came from (a strap id, the "-noop" computed sibling, "apple-health", or "health-connect").
- *  The readings TABLE and the "N readings" header both derive from this ONE list, so they can never
- *  disagree; the raw source maps to a human label via [provenanceDisplayLabel] — the SAME resolver Today
- *  uses, so we never invent a source vocabulary (task #8). */
-internal data class VitalReading(
-    val day: String,
-    val value: Double,
-    val source: String,
-)
-
-/** #377: merge the three step stores into one per-day series with the SAME precedence as the Today
- *  Steps tile — a REAL on-device count ([real], WHOOP 5/MG @57 → DailyMetric.steps) wins, else an
- *  [imported] Health Connect / Apple Health count, else the motion-model [est] (`steps_est`). The three
- *  are disjoint stores so the `?:` chain never double-counts. Ascending by day. Pure for testability. */
-internal fun mergeStepsReadings(
-    real: Map<String, VitalReading>,
-    imported: Map<String, VitalReading>,
-    est: Map<String, VitalReading>,
-): List<VitalReading> =
-    (real.keys + imported.keys + est.keys).toSortedSet()
-        .mapNotNull { d -> real[d] ?: imported[d] ?: est[d] }
-
-/** #616: per-day precedence merge for a metric with disjoint stores (first non-null per day wins),
- *  ascending. The N-store generalisation of [mergeStepsReadings]; calories reuse it as the two-store
- *  on-device (`activeKcalEst`) ?: imported (Apple/Health-Connect `activeKcal`) union. Pure for testability. */
-internal fun mergeReadings(vararg stores: Map<String, VitalReading>): List<VitalReading> =
-    stores.flatMap { it.keys }.toSortedSet()
-        .mapNotNull { day -> stores.firstNotNullOfOrNull { it[day] } }
 
 private data class VitalDetailModel(
     val key: String,
@@ -2201,50 +1855,6 @@ fun VitalDetailScreen(vm: AppViewModel, key: String) {
     }
 }
 
-/** The rows of a vital detail's readings table: each reading's day (localized), its formatted value with
- *  unit, and a human source label. Plain strings so the composable is a thin renderer and the projection
- *  stays unit-testable. */
-internal data class VitalReadingRow(
-    val time: String,
-    val value: String,
-    val source: String,
-)
-
-/**
- * Project a vital's windowed [readings] into table rows, NEWEST FIRST — the same list (so the same count)
- * the "N readings" header shows, guaranteeing the two never drift. Each row pairs the reading's DAY (these
- * vital series carry one aggregated reading per night, so a row's "time" is its calendar date, localized;
- * the date always shows since a charted window spans 2+ days) with the model's own [format]ted value +
- * [unit] and the source label resolved by [provenanceDisplayLabel] — no new source vocabulary (a strap id
- * → "Whoop", its "-noop" sibling → "On-device", "apple-health" → "Apple Health", "health-connect" →
- * "Health Connect"). [strapDeviceId] is the active strap id the label resolver needs.
- */
-internal fun vitalReadingRows(
-    readings: List<VitalReading>,
-    unit: String,
-    strapDeviceId: String,
-    format: (Double) -> String,
-): List<VitalReadingRow> =
-    readings.asReversed().map { reading ->
-        VitalReadingRow(
-            time = vitalReadingDateLabel(reading.day),
-            value = "${format(reading.value)} $unit".trim(),
-            source = provenanceDisplayLabel(reading.source, strapDeviceId),
-        )
-    }
-
-/** "9 Jun" for a "YYYY-MM-DD" reading day (today / yesterday read as words to match the hero "as of"
- *  line); the verbatim string if it doesn't parse. Locale.US month, matching [asOfLabel]. */
-internal fun vitalReadingDateLabel(day: String): String {
-    val date = runCatching { LocalDate.parse(day) }.getOrNull() ?: return day
-    val today = LocalDate.now()
-    return when (date) {
-        today -> "Today"
-        today.minusDays(1) -> "Yesterday"
-        else -> date.format(DateTimeFormatter.ofPattern("d MMM", Locale.US))
-    }
-}
-
 /** The readings table below a vital's chart: one row per windowed reading (newest first), each showing
  *  its day, formatted value, and source (tinted by [provenanceLabelTint], so the same source reads the
  *  same colour as the Today rings). Empty [rows] render nothing. */
@@ -2319,144 +1929,6 @@ private fun VitalReadingsTable(rows: List<VitalReadingRow>) {
 @Composable
 private fun RecentDaySelectorBar(selectedOffset: Int, onSelect: (Int) -> Unit) {
     ThreeDaySelectorBar(selectedOffset = selectedOffset, onSelect = onSelect)
-}
-
-private fun latestVitals(days: List<DailyMetric>, tempUnit: TemperatureUnit): List<Vital> {
-    val emptyByKey = vitalsFor(null, days, tempUnit).associateBy { it.key }
-    return listOf(
-        latestVital("resp", days, tempUnit, emptyByKey) { it.respRateBpm != null },
-        latestVital("spo2", days, tempUnit, emptyByKey) { it.spo2Pct != null },
-        latestVital("spo2raw", days, tempUnit, emptyByKey) { it.spo2Red != null && it.spo2Ir != null },
-        latestVital("rhr", days, tempUnit, emptyByKey) { it.restingHr != null },
-        latestVital("hrv", days, tempUnit, emptyByKey) { it.avgHrv != null },
-        latestVital("skin", days, tempUnit, emptyByKey) { it.skinTempDevC != null },
-    )
-}
-
-private fun latestVital(
-    key: String,
-    days: List<DailyMetric>,
-    tempUnit: TemperatureUnit,
-    emptyByKey: Map<String, Vital>,
-    hasValue: (DailyMetric) -> Boolean,
-): Vital {
-    val row = days.asReversed().firstOrNull(hasValue)
-    return row
-        ?.let { latestRow -> vitalsFor(latestRow, days, tempUnit).firstOrNull { it.key == key } }
-        ?.copy(asOfLabel = asOfLabel(row.day))
-        ?: emptyByKey.getValue(key)
-}
-
-private fun selectedDayLabel(offset: Int): String = when (offset) {
-    0 -> "Today"
-    1 -> "Yesterday"
-    else -> "2 days ago"
-}
-
-private fun missingVitalsTitle(offset: Int): String = when (offset) {
-    0 -> "We didn't get today's data"
-    1 -> "We didn't get yesterday's data"
-    else -> "We didn't get data from 2 days ago"
-}
-
-private fun asOfLabel(day: String?): String? {
-    if (day.isNullOrBlank()) return null
-    val date = runCatching { LocalDate.parse(day) }.getOrNull() ?: return "as of $day"
-    val today = LocalDate.now()
-    return when (date) {
-        today -> "as of today"
-        today.minusDays(1) -> "as of yesterday"
-        else -> "as of ${date.format(DateTimeFormatter.ofPattern("d MMM", Locale.US))}"
-    }
-}
-
-internal enum class VitalDetailRange(val label: String, val days: Long?) {
-    WEEK("W", 7),
-    TWO_WEEK("2W", 14),
-    THREE_WEEK("3W", 21),
-    MONTH("M", 30),
-    THREE_MONTH("3M", 90),
-    SIX_MONTH("6M", 180),
-    YEAR("1Y", 365),
-    ALL("ALL", null),
-}
-
-/** Days spanned by a vital's history: last point's day minus first point's day in epoch days (0 for
- *  a single day or unparseable bounds). Points arrive oldest-first from buildVitalDetail. */
-internal fun vitalHistorySpanDays(points: List<Pair<String, Double>>): Long {
-    val first = points.firstOrNull()?.first?.let { runCatching { LocalDate.parse(it) }.getOrNull() } ?: return 0L
-    val last = points.lastOrNull()?.first?.let { runCatching { LocalDate.parse(it) }.getOrNull() } ?: return 0L
-    return (last.toEpochDay() - first.toEpochDay()).coerceAtLeast(0L)
-}
-
-/** #943 (ryanbr): which range chips have anything NEW to show. filterVitalPoints windows off the
- *  LATEST reading, so with under a week of history every window returned the identical full point set
- *  and all six chips drew the same line (a week of data stretched full-width under a "1Y" label). A
- *  range only differs from its predecessor once the data span EXCEEDS the predecessor's window, so the
- *  unlocked set is a contiguous prefix: W always, 2W once span > 7 days, 3W once > 14, M once > 21,
- *  3M once > 30, 6M once > 90, 1Y once > 180, ALL once > 365. (The 1D/2D experiment was dropped: daily
- *  metrics hold at most one point per day, so those windows could never draw a line.) Locked chips render
- *  disabled rather than hidden so a calibrating user still learns the longer views exist; W (the shortest)
- *  staying unconditional means nobody is ever stranded with zero ranges. */
-/**
- * The range the chips + caption actually describe, resolved NON-DESTRUCTIVELY (Swift parity with
- * MetricExplorerView.coercedSelection). A locked selection renders as the largest unlocked range with
- * a real finite window that is <= the selection, else WEEK. NOT ALL: coercing a locked default to ALL
- * would jump a calibrating user to the everything view. An unlocked selection is used verbatim, so the
- * chip un-coerces on its own once history grows.
- */
-internal fun coercedVitalRange(range: VitalDetailRange, unlocked: List<VitalDetailRange>): VitalDetailRange {
-    if (range in unlocked) return range
-    return VitalDetailRange.entries
-        .filter { it.days != null && it.ordinal <= range.ordinal && it in unlocked }
-        .maxByOrNull { it.ordinal }
-        ?: VitalDetailRange.WEEK
-}
-
-internal fun unlockedVitalRanges(spanDays: Long): List<VitalDetailRange> {
-    val ranges = VitalDetailRange.entries
-    val unlocked = mutableListOf(ranges.first())
-    for (i in 1 until ranges.size) {
-        val previousWindow = ranges[i - 1].days ?: break
-        if (spanDays > previousWindow) unlocked += ranges[i] else break
-    }
-    // ALL is never gated (Swift parity): a calibrating user can always see their full history,
-    // even when it happens to draw the same points as a shorter window.
-    val all = ranges.last()
-    if (all.days == null && all !in unlocked) unlocked += all
-    return unlocked
-}
-
-internal fun filterVitalPoints(
-    points: List<Pair<String, Double>>,
-    range: VitalDetailRange,
-): List<Pair<String, Double>> {
-    val windowDays = range.days ?: return points
-    val latestDate = points.lastOrNull()?.first?.let { runCatching { LocalDate.parse(it) }.getOrNull() }
-        ?: return points.takeLast(windowDays.toInt())
-    val cutoff = latestDate.minusDays(windowDays - 1)
-    val filtered = points.filter { (day, _) ->
-        runCatching { LocalDate.parse(day) }.getOrNull()?.let { !it.isBefore(cutoff) } ?: false
-    }
-    return filtered.ifEmpty { points.takeLast(windowDays.toInt()) }
-}
-
-/** [filterVitalPoints] for the source-carrying [VitalReading] list — the SAME latest-relative window, so
- *  the readings table and the chart always agree on which readings are in view (task #8). Kept as a twin
- *  of the point filter (identical windowing) rather than shared-generic to preserve the pinned-test shape
- *  of [filterVitalPoints]. */
-internal fun filterVitalReadings(
-    readings: List<VitalReading>,
-    range: VitalDetailRange,
-): List<VitalReading> {
-    val windowDays = range.days ?: return readings
-    val latestDate = readings.lastOrNull()?.day?.let { runCatching { LocalDate.parse(it) }.getOrNull() }
-        ?: return readings.takeLast(windowDays.toInt())
-    val cutoff = latestDate.minusDays(windowDays - 1)
-    val filtered = readings.filter { reading ->
-        runCatching { LocalDate.parse(reading.day) }.getOrNull()?.let { !it.isBefore(cutoff) } ?: false
-    }
-    return filtered.ifEmpty { readings.takeLast(windowDays.toInt()) }
 }
 
 private fun buildVitalDetail(
